move some stream rpc client tests to new file
there were multiple describe sections in one file, too huge, too confusing.
